+++ src/easydiffraction/Profiles/common.py
@@ -127,8 +127,8 @@ class _PowderBase(BaseObj):
             name, _phases=phases, _parameters=parameters, _pattern=pattern
         )
 
-        from easydiffraction.Interfaces.types import Neutron
-        from easydiffraction.Interfaces.types import Powder
+        from easydiffraction.calculators.wrapper_types import Neutron
+        from easydiffraction.calculators.wrapper_types import Powder
         from easydiffraction.Profiles.P1D import Instrument1DCWParameters
         from easydiffraction.Profiles.P1D import Instrument1DTOFParameters
         from easydiffraction.Profiles.P1D import PolPowder1DParameters as Pattern1D_Pol
@@ -139,19 +139,19 @@ class _PowderBase(BaseObj):
         self._update_bases(Neutron)
 
         if getattr(pattern, "__old_class__", pattern.__class__) == Pattern1D:
-            from easydiffraction.Interfaces.types import UPol
+            from easydiffraction.calculators.wrapper_types import UPol
 
             self._update_bases(UPol)
         elif getattr(pattern, "__old_class__", pattern.__class__) == Pattern1D_Pol:
-            from easydiffraction.Interfaces.types import Pol
+            from easydiffraction.calculators.wrapper_types import Pol
 
             self._update_bases(Pol)
         if isinstance(parameters, Instrument1DCWParameters):
-            from easydiffraction.Interfaces.types import CW
+            from easydiffraction.calculators.wrapper_types import CW
 
             self._update_bases(CW)
         elif isinstance(parameters, Instrument1DTOFParameters):
-            from easydiffraction.Interfaces.types import TOF
+            from easydiffraction.calculators.wrapper_types import TOF
 
             self._update_bases(TOF)
